Added rhyme checking
Now challenge.c can check if the last word of the input phrase matches a word in rhymelist.txt between the proper lines and return 1 if the word rhymes and return 0 if it does not. In the sample rhymelist, I've put some words that rhyme with 3 words. The program only checks if the input word rhymes with "cool" by checking the lines 13 through 17.